x86_64 pointer typecast fix by (Martin Drab <drab kepler.fjfi.cvut cz>)
Originally committed as revision 3845 to svn://svn.ffmpeg.org/ffmpeg/trunk

+++ libavcodec/h263.c
@@ -3009,7 +3009,7 @@ void ff_mpeg4_init_partitions(MpegEncContext *s)
     uint8_t *start= pbBufPtr(&s->pb);
     uint8_t *end= s->pb.buf_end;
     int size= end - start;
-    int pb_size = (((int)start + size/3)&(~3)) - (int)start;
+    int pb_size = (((long)start + size/3)&(~3)) - (long)start;
     int tex_size= (size - 2*pb_size)&(~3);
     
     set_put_bits_buffer_size(&s->pb, pb_size);
